Replace WI
Replacing with an old WI dataset as triage. The vote counts in the dataset that is currently in this repository are wrong (sometimes by as much as 25% off for major offices) and the reason is not immediately obvious. This file is not completely cleaned, but it passes checks for vote count accuracy in major offices. We will update the format to match the other state files as soon as we have the chance